backend/requirements.txt
flask
flask-cors
litellm
python-dotenv
requests
google-genai>=1.9.0
pillow>=10.0.0
